Modified Quest Method checkQuest
In class Quest and its subclasses, I have copied the checkQuest method and changed it to only take in a dictionary regardless of Quest type. This allows for modularity in the future. The previous versions of checkQuest are still available just in case, but should be commented out later. I have also modded the comments in ViewController to show what the dictionary must contain per game to successfully use checkQuest, as well as an example in the "breathe focus" if-case.