Updated Quotes, Earned Pet Names
The quotes list has been updated with all the quotes provided except two that were too large to be usable.

When the Daily Quest rewards a Pet, the Pet's name is now properly capitalized.

Tested Daily Quests to ensure that the player cannot reroll once a Quest is completed and Quests and rerolls properly reset upon Quest expiration.